Apple updates

+++ util/Character.h
@@ -12,7 +12,7 @@ struct Character {
     GLuint TextureID;   // ID handle of the glyph texture
     glm::ivec2 Size;    // Size of glyph
     glm::ivec2 Bearing;  // Offset from baseline to left/top of glyph
-    GLuint Advance;    // Horizontal offset to advance to next glyph
+    signed long Advance;    // Horizontal offset to advance to next glyph
 };
 
 #endif //GAME_CHARACTER_H
